Refactor timetable utils into smaller files - The timetable utils file was originally 1400+ locs while its tests is another 1400. Refactoring them into smaller logical files makes it significantly easier to read, writes and test - Split up symbols are reexported from utils/timetables
