+++ website/src/utils/timetables/exams.ts
@@ -1,0 +1,113 @@
+import { groupBy } from 'lodash-es';
+
+import { Module, Semester } from 'types/modules';
+
+import { ExamClashes } from 'types/views';
+
+import { getExamDate, getExamDuration } from 'utils/modules';
+
+// Creates a key using only the exam date string (without time)
+export function getExamDateOnly(module: Module, semester: Semester): string | undefined {
+  const examDateTime = getExamDate(module, semester);
+  return examDateTime?.slice(0, 10);
+}
+
+// Returns the start time of the exam as an epoch time (number). Throws an error if the module
+// does not have an exam date.
+function getValidExamStartTimeAsEpoch(module: Module, semester: Semester): number {
+  const startTimeString = getExamDate(module, semester);
+  if (startTimeString === null) {
+    throw new Error('Courses tested for clashes must have exam dates and durations!');
+  }
+  return new Date(startTimeString).getTime();
+}
+
+// Returns the end time of the exam as an epoch time (number). Throws an error if the module
+// does not have an exam date or duration.
+function getValidExamEndTimeAsEpoch(module: Module, semester: Semester): number {
+  const duration = getExamDuration(module, semester);
+  if (duration === null) {
+    throw new Error('Courses tested for clashes must have exam dates and durations!');
+  }
+  const startEpoch = getValidExamStartTimeAsEpoch(module, semester);
+  return startEpoch + duration * 60 * 1000;
+}
+
+// Find all exam clashes between modules in semester
+// Returns object associating exam dates with the modules clashing on those dates
+export function findExamClashes(modules: Module[], semester: Semester): ExamClashes {
+  // Filter away modules without exam dates or exam durations
+  const filteredModules = modules.filter(
+    (module) =>
+      getExamDate(module, semester) !== null && getExamDuration(module, semester) !== null,
+  );
+
+  const groupedModules = groupBy(filteredModules, (module) => getExamDateOnly(module, semester));
+
+  const clashes: ExamClashes = {};
+
+  Object.values(groupedModules).forEach((sameDayMods) => {
+    // Sort sameDayMods by exam start time
+    sameDayMods.sort((a, b) => {
+      const aStartEpoch = getValidExamStartTimeAsEpoch(a, semester);
+      const bStartEpoch = getValidExamStartTimeAsEpoch(b, semester);
+
+      // Use end time as secondary key
+      const aEndEpoch = getValidExamEndTimeAsEpoch(a, semester);
+      const bEndEpoch = getValidExamEndTimeAsEpoch(b, semester);
+
+      if (aStartEpoch === bStartEpoch) {
+        return aEndEpoch - bEndEpoch;
+      }
+
+      return aStartEpoch - bStartEpoch;
+    });
+
+    // Initialize an empty list to hold the groups of overlapping intervals
+    // Each group will itself be a list of intervals
+    const overlappingGroups: Module[][] = [];
+
+    let currentOverlapEnd = 0;
+    let currentOverlappingMods: Module[] = [];
+
+    sameDayMods.forEach((mod, modIndex) => {
+      if (modIndex > 0 && getValidExamStartTimeAsEpoch(mod, semester) < currentOverlapEnd) {
+        currentOverlappingMods.push(mod);
+      } else {
+        // The current course does not overlap with the current group, so we reset
+        // the current group and start a new one
+        if (currentOverlappingMods.length > 1) {
+          // If the current group has more than one module, we add it to the list of clashes
+          overlappingGroups.push(currentOverlappingMods);
+        }
+        currentOverlapEnd = getValidExamEndTimeAsEpoch(mod, semester);
+        currentOverlappingMods = [mod];
+      }
+    });
+
+    // Add the last group to the list of clashes if applicable
+    if (currentOverlappingMods.length > 1) {
+      overlappingGroups.push(currentOverlappingMods);
+    }
+
+    overlappingGroups.forEach((group) => {
+      // Displayed clashing date and time, which is the start time of the last module in the group
+      const clashingDateTime = getExamDate(group[group.length - 1], semester);
+
+      if (clashingDateTime === null) {
+        throw new Error('Courses tested for clashes must have exam dates and durations!');
+      }
+
+      // Populate the clashes object to be returned
+      group.forEach((mod) => {
+        if (!clashes[clashingDateTime]) {
+          clashes[clashingDateTime] = [mod];
+        } else {
+          clashes[clashingDateTime].push(mod);
+        }
+      });
+    });
+  });
+
+  return clashes;
+}
